% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BM_from_XML.R
\name{BM_addAttributes}
\alias{BM_addAttributes}
\title{Add Query Attributes To biomaRt Call}
\usage{
BM_addAttributes(cll, atts)
}
\arguments{
\item{cll}{biomaRt call as returned by format_BM_from_XML()}

\item{atts}{String vector of names of attributes to add.}
}
\value{
The updated biomaRt call object
}
\description{
Add a set of attributes (output fields) requested by the query.
}
