% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BM_from_XML.R
\name{BM_addFilters}
\alias{BM_addFilters}
\title{Add filters to a biomaRt call.}
\usage{
BM_addFilters(cll, ...)
}
\arguments{
\item{cll}{biomaRt call as returned by format_BM_from_XML()}

\item{...}{key=value arguments to add to the biomaRt call.}
}
\value{
The updated biomaRt call object
}
\description{
Add a set of filters to a biomaRt call.
}
\seealso{
\code{\link{BM_removeFilters}}, \code{\link{BM_substituteFilterValues}}
}
