% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BM_from_XML.R
\name{BM_removeAttributes}
\alias{BM_removeAttributes}
\title{Remove Query Attributes From biomaRt Call}
\usage{
BM_removeAttributes(cll, atts)
}
\arguments{
\item{cll}{biomaRt call as returned by format_BM_from_XML()}

\item{atts}{String vector of names of attributes to remove.}
}
\value{
The updated biomaRt call object
}
\description{
Remove a set of attributes (output fields) requested by the query.
}
