% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BM_from_XML.R
\name{BM_removeFilters}
\alias{BM_removeFilters}
\title{Remove Filters From a biomaRt Call.}
\usage{
BM_removeFilters(cll, removefilter)
}
\value{
The updated biomaRt call object
}
\description{
Remove filters and their values from the biomaRt call.
}
\seealso{
\code{\link{BM_addFilters}}, \code{\link{BM_substituteFilterValues}}
}
