% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BM_from_XML.R
\name{BM_substituteFilterValues}
\alias{BM_substituteFilterValues}
\title{Override filters in the biomaRt call.}
\usage{
BM_substituteFilterValues(cll, ...)
}
\arguments{
\item{cll}{biomaRt call as returned by format_BM_from_XML()}

\item{...}{key=value arguments to replace in the biomaRt call.}
}
\value{
The updated biomaRt call object
}
\description{
Replace filter values in a biomaRt call.
}
\seealso{
\code{\link{BM_removeFilters}}, \code{\link{BM_addFilters}}
}
