% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BM_from_XML.R
\name{format_BM_from_XML}
\alias{format_BM_from_XML}
\title{Translate XML Query to R Code}
\usage{
format_BM_from_XML(xmltext, printCode = TRUE)
}
\arguments{
\item{xmltext}{XML text copied from \url{https://parasite.wormbase.org/biomart}}

\item{printCode}{TRUE/FALSE Print the R code to the console in addition to returning the call.}
}
\value{
A 'call' object that can be run with runWithMart().
}
\description{
Take a text object returned from the XML formatted output at
\url{https://parasite.wormbase.org/biomart} and return a function call
to biomaRt::getBM with the filter names, values, and requested output
fields.
}
\examples{
# A query that extracts all C. elegans long noncoding RNAs.
query = format_BM_from_XML('<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE Query>
<Query  virtualSchemaName = "parasite_mart" formatter = "TSV" header = "0" uniqueRows = "0" count = "" datasetConfigVersion = "0.6" >
  <Dataset name = "wbps_gene" interface = "default" >
     <Filter name = "biotype" value = "lincRNA"/>
     <Filter name = "species_id_1010" value = "caelegprjna13758"/>
     <Attribute name = "wbps_gene_id" />
  </Dataset>
</Query>')

mart = getParamart()
lincs = runWithMart(query, mart)
head(lincs)
}
\seealso{
\code{\link[biomaRt]{getBM}}, \code{\link{runWithMart}}
}
