% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BM_from_XML.R
\name{runWithMart}
\alias{runWithMart}
\title{Run a pre-formatted query with the provided "Mart" object.}
\usage{
runWithMart(cll, mart)
}
\arguments{
\item{cll}{biomaRt call as returned by format_BM_from_XML()}

\item{mart}{a "Mart" object as returned by getParamart() or useMart().}
}
\value{
The query result as a dataframe.
}
\description{
Run a pre-formatted query, as returned by format_BM_from_XML() to biomaRt
with a "Mart" object as returned by getParamart() or useMart().
}
\examples{
query = format_BM_from_XML('<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE Query>
<Query  virtualSchemaName = "parasite_mart" formatter = "TSV" header = "0" uniqueRows = "0" count = "" datasetConfigVersion = "0.6" >
  <Dataset name = "wbps_gene" interface = "default" >
     <Filter name = "biotype" value = "lincRNA"/>
     <Filter name = "species_id_1010" value = "caelegprjna13758"/>
     <Attribute name = "wbps_gene_id" />
  </Dataset>
</Query>')

mart = getParamart()
lincs = runWithMart(query, mart)
}
\seealso{
\code{\link[biomaRt]{getBM}}, \code{\link{getParamart}}
}
